refactor(lockfile): split pnpm modules
