+++ crates/aube-lockfile/src/pnpm/mod.rs
@@ -1,0 +1,11 @@
+mod dep_path;
+mod format;
+mod raw;
+mod read;
+mod write;
+
+#[cfg(test)]
+mod tests;
+
+pub use read::parse;
+pub use write::write;
